Disabled a couple of the statistical tests.
They have already served their purpose in verifying my understanding of what
was going on, but they are sufficiently slow and flakey that it isn't worth
running them every time.  (Admittedly, the flakiness is probably revealing
that the metropolis algorithm doesn't do the best job in the world of
effectively sampling the distribution, but I'm deeming the results to mean
that it's "good enough" --- until I think of a better way to sample paths. :-) )