Fixed space leaks.
The price paid for this is that I now force computation of the particle
separations.  I had to choose between this option and having them be re-computed
entirely upon demand, rather than only re-computing the slices that changed
since the last configuration.

I also fixed a space leak in the observable updated;  it now forces evaluation
of the double.